Skip building if nightly item already there

+++ tools/nightly22/build-nightly.sh
@@ -27,25 +27,32 @@ docker rmi -f new-${docker}
 code/tools/nightly22/build-sdcard.sh
 
 # Build Linux companion
-docker run -dit --name companion -v /home/opentx/${docker}:/opentx ${docker}
-docker exec companion sh -c "mkdir -p build && cd build && cmake /opentx/code && cp radio/src/stamp.h /opentx/binaries/stamp-opentx.txt"
-docker exec companion rm -rf build
-docker exec companion /opentx/code/tools/build-companion.sh /opentx/code /opentx/binaries/ ${suffix}
-docker stop companion
-docker rm companion
-cp -f  binaries/*.deb ${output}/companion/linux/companion22_${version}${suffix}_amd64.deb
+if [ ! -f ${output}/companion/linux/companion22_${version}${suffix}_amd64.deb]; then
+  docker run -dit --name companion -v /home/opentx/${docker}:/opentx ${docker}
+  docker exec companion sh -c "mkdir -p build && cd build && cmake /opentx/code && cp radio/src/stamp.h /opentx/binaries/stamp-opentx.txt"
+  docker exec companion rm -rf build
+  docker exec companion /opentx/code/tools/build-companion.sh /opentx/code /opentx/binaries/ ${suffix}
+  docker stop companion
+  docker rm companion
+  cp -f  binaries/*.deb ${output}/companion/linux/companion22_${version}${suffix}_amd64.deb
+fi
 
 # Request companion compilation on Windows
-cd ${output}/companion/windows
-wget -qO- http://winbox.open-tx.org/companion-builds/compile22.php?branch=$branch\&suffix=${suffix}
-wget -O companion-windows-${version}${suffix}.exe http://winbox.open-tx.org/companion-builds/companion-windows-${version}${suffix}.exe
-chmod -Rf g+w companion-windows-${version}${suffix}.exe
+if [ ! -f ${output}/companion/windows/companion-windows-${version}${suffix}.exe]; then
+  cd ${output}/companion/windows
+  wget -qO- http://winbox.open-tx.org/companion-builds/compile22.php?branch=$branch\&suffix=${suffix}
+  wget -O companion-windows-${version}${suffix}.exe http://winbox.open-tx.org/companion-builds/companion-windows-${version}${suffix}.exe
+  chmod -Rf g+w companion-windows-${version}${suffix}.exe
+fi
+
 
 # Request companion compilation on Mac OS X
-cd ${output}/companion/macosx
-wget -qO- http://opentx.blinkt.de:8080/~opentx/build-opentx.py?branch=${branch}\&suffix=${suffix}
-wget -O opentx-companion-${version}${suffix}.dmg http://opentx.blinkt.de:8080/~opentx/builds/opentx-companion-${version}${suffix}.dmg
-chmod -Rf g+w opentx-companion-${version}${suffix}.dmg
+if [ ! -f ${output}/companion/macosx/opentx-companion-${version}${suffix}.dmg]; then
+  cd ${output}/companion/macosx
+  wget -qO- http://opentx.blinkt.de:8080/~opentx/build-opentx.py?branch=${branch}\&suffix=${suffix}
+  wget -O opentx-companion-${version}${suffix}.dmg http://opentx.blinkt.de:8080/~opentx/builds/opentx-companion-${version}${suffix}.dmg
+  chmod -Rf g+w opentx-companion-${version}${suffix}.dmg
+fi
 
 # Update stamps
 cp -f  $workdir/binaries/stamp-opentx.txt ${output}/firmware
